+++ layers/Engine/packages/engine-wp/src/HelperServices/TemplateHelpers.php
@@ -32,13 +32,17 @@ class TemplateHelpers implements TemplateHelpersInterface
     }
 
     /**
-     * Add a hook to send the Response to the client
+     * Add a hook to send the Response to the client.
+     *
+     * `App::isInitialized()` is needed to avoid conflicts with other plugins,
+     * such as the "All In One Security" plugin.
+     * @see https://github.com/GatoGraphQL/GatoGraphQL/issues/3283
      */
     public function sendResponseToClient(): void
     {
         \add_filter(
             'template_include',
-            fn (?string $template) => $this->getSendResponseTemplateFile(),
+            fn (?string $template) => !App::isInitialized() ? $template : $this->getSendResponseTemplateFile(),
             PHP_INT_MAX // Execute last
         );
     }