+++ layers/GatoGraphQLForWP/phpunit-plugins/gatographql-testing/src/Webserver/LandoAdapter.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace PHPUnitForGatoGraphQL\GatoGraphQLTesting\Webserver;
 
 use GatoGraphQL\GatoGraphQL\Facades\Request\PrematureRequestServiceFacade;
+use GatoGraphQL\GatoGraphQL\Services\Helpers\EndpointHelpers;
 use PoP\ComponentModel\App;
 use PoP\ComponentModel\Engine\EngineHookNames;
+use PoP\Root\Facades\Instances\SystemInstanceManagerFacade;
 
 use function add_filter;
 use function remove_filter;
@@ -14,10 +16,9 @@ use function remove_filter;
 /**
  * When not using Lando with a proxy, the assigned URL
  * will be something like "localhost:54023", however,
- * this URL is not accessible from within the container,
- * as the port is not mapped.
+ * this URL is not accessible from within the container.
  *
- * Hence, convert this into "localhost".
+ * Convert it into "host.docker.internal:54023", then it works.
  */
 class LandoAdapter
 {
@@ -27,7 +28,12 @@ class LandoAdapter
             'init',
             function (): void {
                 $applicationStateHelperService = PrematureRequestServiceFacade::getInstance();
-                if (!$applicationStateHelperService->isPubliclyExposedGraphQLAPIRequest()) {
+                /** @var EndpointHelpers */
+                $endpointHelpers = SystemInstanceManagerFacade::getInstance()->getInstance(EndpointHelpers::class);
+                if (
+                    !$applicationStateHelperService->isPubliclyExposedGraphQLAPIRequest()
+                    && !$endpointHelpers->isRequestingAdminGraphQLEndpoint()
+                ) {
                     return;
                 }
 
@@ -61,11 +67,11 @@ class LandoAdapter
      */
     public function maybeRemovePortFromLocalhostURL(string $url): string
     {
-        if (str_starts_with($url, 'https://localhost:')) {
-            return 'https://localhost';
-        }
-        if (str_starts_with($url, 'http://localhost:')) {
-            return 'http://localhost';
+        if (\preg_match('#^(https?)://localhost:(\d+)(/.*)?$#', $url, $matches) === 1) {
+            $scheme = $matches[1];
+            $port = $matches[2];
+            $path = $matches[3] ?? '';
+            return \sprintf('%s://host.docker.internal:%s%s', $scheme, $port, $path);
         }
         return $url;
     }