+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 17.1.0 - 23/03/2026
+
+### Improvements
+
+- Updated Translation extension docs on using a self-hosted LLM (#3282)
+
+### Fixed
+
+- Conflict with the "All In One Security" plugin (#3283)
+- JetEngine: Fix missing images in docs (#3279)
+
 ## 17.0.0 - 03/03/2026
 
 ### Breaking changes